requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
paho-mqtt==1.6.1
PyYAML==6.0
telepot==12.7
urllib3==1.26.9
yarl==1.7.2
